move ValidatorNodeProvider from client.rs to node.rs
